rename LocalStorage hook to useLocalStorage
